templates/ddd-service-native-db/pyproject.toml
[tool.poetry]
name = "${PROJECT_NAME}"
version = "${PROJECT_VERSION}"
description = "${PROJECT_DESCRIPTION}"
authors = []
readme = "README.md"
license = "MIT"
homepage = "${HOMEPAGE}"
repository = "${REPOSITORY}"
keywords = ["python"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[tool.poetry.urls]
"Bug Reports" = "${BUG_REPORTS_URL}"
"Source" = "${SOURCE_URL}"

[tool.poetry.dependencies]
python = ">=3.10"
python-dotenv = ">=1.0.0"
joblib = ">=1.3"
pyodbc = ">=5.0"
oracledb = ">=2.0"
psycopg = ">=3.1"
mysql-connector-python = ">=8.3"
stpstone = ">=3.0.0"

[tool.poetry.group.dev.dependencies]
pre-commit = ">=4.1.0"
pretty-errors = ">=1.2.25"
commitizen = ">=4.4.1"
poetry-plugin-export = ">=1.8"

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"
